moved web3 to getweb3 file
