+++ client/src/Routes.js
@@ -2,6 +2,7 @@ import React from 'react'
 import { BrowserRouter as Router, Route, Switch } from 'react-router-dom';
 import Home from './components/Home'
 import Nav from './components/Nav'
+import "./App.css";
 
 const Routes = () => {
   return (
