+++ client/src/components/Wallet.js
@@ -1,28 +1,5 @@
 import React from 'react'
-import Nav from './Nav'
-import Torus from "@toruslabs/torus-embed";
-import Web3 from "web3";
-
-const web3Obj = {
-  web3: new Web3(),
-  setweb3: function(provider) {
-    const web3Inst = new Web3(provider)
-    web3Obj.web3 = web3Inst
-    sessionStorage.setItem('pageUsingTorus', true)
-  },
-  initialize: async function() {
-    const torus = new Torus()
-    await torus.init({
-      network: {
-        host: 'HTTP://127.0.0.1:7545',
-        networkName: 'dev'
-      },
-      enableLogging: false
-    })
-    await torus.login()
-    web3Obj.setweb3(torus.provider)
-  }
-}
+import web3Obj from '../getWeb3'
 
 class Wallet extends React.Component {
   constructor() {
@@ -68,11 +45,12 @@ class Wallet extends React.Component {
       <div class="wallet">
           {
             this.state.account ? 
+            
             <div class="accountInfo">
               <p>Account: {this.state.account.slice(0, 5)}</p>
               <p>Balance: {this.state.balance}</p>
             </div>
-          
+  
             : 
             
             <div>
